+++ docs/main/changelog.rst
@@ -12,6 +12,7 @@ Development version
 - Add ``Configuration.from_yaml()`` method to load configuration from yaml file.
 - Add ``Configuration.from_ini()`` method to load configuration from ini file.
 - Add ``Configuration.from_dict()`` method to load configuration from dictionary.
+- Add default value for ``name`` argument of ``Configuration`` provider.
 - Remove undocumented positional parameter of ``DependenciesContainer`` provider.
 
 3.17.1